BS 174 | Vijay, Mani | refactored the code to a package
